feat: 계층형 캐시 도입 및 Redis 안정성 강화
- Caffeine(L1)과 Redis(L2)를 조합한 다중 계층 캐시 구조 구현
- Jackson 역직렬화 시 Integer 캐스팅 에러 방지를 위해 반환 타입을 Double로 변경
- Docker-compose 내 Redis maxmemory 제한 및 LRU 방출 정책 설정 추가
- AOP 프록시 환경에서의 동시성 테스트 코드 정상화

+++ src/main/java/maple/expectation/service/v2/CubeTrialsProvider.java
@@ -7,5 +7,5 @@ public interface CubeTrialsProvider {
     /**
      * 특정 큐브와 설정값에 따른 목표 옵션 도달 기대 시도 횟수를 반환합니다.
      */
-    Long calculateExpectedTrials(CubeCalculationInput input, CubeType type);
+    Double calculateExpectedTrials(CubeCalculationInput input, CubeType type);
 }
